handle protected mmcblk_boot_ devices
Some block devices support physical write protection. The kernel
provides a standard interface to enable or disable protection in
/sys/class/block/*/force_ro.

This patch adds functionality to automatically detect these memory
types. If read-only mode is enabled on the partition on which the
uboot environment must be written, libubootenv temporarily switches
to read/write mode.ease enter the commit message for your changes.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>