uboot_env: fix the resarch of ubi volume
In the function ubi_get_vol_id, when we search the volume id
with the volume name, we try all id from 0 to num_volume. But
the volume id may be greater than the actual number of volume.
To avoid this issue, we try all the id from 0 to 128 (max ubi
volume id).

Signed-off-by: Philippe Reynes <philippe.reynes@softathome.com>