ubi, write: fix invalid envsize ptr to UBI_IOCVOLUP
The UBI_IOCVOLUP ioctl expects a __s64*, however the struct
uboot_flash_env.envsize is of type size_t, which depending on the
architecture may be only 4 bytes long and cause the ioctl to read a
bogus size value. This fixes this by converting the size_t envsize to
a int64_t and passing a ptr to that to the ioctl.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>
Acked-by: Stefano Babic <sbabic@denx.de>