Force writing of environment if default is used
Before commit ed1a53ec "Dont store to device if no value changes" it was
possible to fill an uninitialized or broken environment using the
command "fw_setenv -f <default-environment-file>". This was considered
to be a feature.

The problem occurs because fw_setenv will now skip writing the
environment in case no environment variables to be changed are passed to
fw_setenv.

This commit will fix this problem by always storing the environment if
loading the environment failed and a default environment was given.

Signed-off-by: Wang Xin <xin.wang7@cn.bosch.com>
Signed-off-by: Mark Jonas <mark.jonas@de.bosch.com>
Acked-by: Stefano Babic <sbabic@denx.de>