Always fsync file writes
Currently fsync() only gets called at the end of fileprotect(), which
only occurs on mmcblk devices. I am storing a U-Boot environment inside
a fat32 filesystem, so the fsync() call does not get hit, and so the new
environment may be lost if power is lost.

Instead always fsync after the write has completed, regardless of the
device type.

Signed-off-by: Brandon Maier <brandon.maier@rockwellcollins.com>
Acked-by: Stefano Babic <sbabic@denx.de>