Add new configuration file in YAML
Legacy configuration file was taken by U-Boot project. Its format is
very simple, but unflexible and it cannot be extended. Changes in the
format could lead to uncompatibility with U-Boot tools, that are still
part of U-Boot.

This introduces a new format, letting the old format as fallback if YAML
cannot be parsed. It uses the libyaml library.

Multiple sets can be configured - an environment is not bound with the
bootloader, and can be used for other purposes, but still having the
features provided by the handling for the bootloader (redundancy,
power-cut safe).

See documentation for the format of the yaml file. Each not recognized
keyword generates an error.

Signed-off-by: Stefano Babic <sbabic@denx.de>