libuboot_read_multiple_config: clamp return value
According to the documentation, the function returns zero on success,
else a negative value.

However, at the moment the return value of parse_yaml_config might
also be passed through which can be FAILURE (define as 1).

So map this one to a generic -1 and pass all other as-is
to match the documented API.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>
Acked-by: Stefano Babic <sbabic@denx.de>