Add function to return the library version
This helps the caller to identify the version of the running version of
libubootenv. It returns a pointer to a static structure. Applications
should use this information to judge if things are possible to do or not,
instead of using compile-time checks, as libraries loaded via dlopen()
can be changed independent from the application.

Signed-off-by: Stefano Babic <sbabic@denx.de>