Allow specifying UBI Volumes by MTD path
In cases where the env is in an unattached UBI volume we may need
to specify the volume name via the MTD path to attach.

For this to work correctly we should first scan for any already
attached UBI devices to see if they correspond to the mtd partition
number in our configuration, if no corresponding attached mtd devices
are found we should attempt to ubiattach to the configured MTD path.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>