Let set a list of writable variables
In U-Boot, it is possible to list which variables could be changed via
CONFIG_ENV_WRITEABLE_LIST. This was not supported by this library, but
it does not raise a security leak. In fact, U-Boot will simply discard
if a variable is not in the list. However, in user space any variable
can be set, and the output of fw_printenv is not what the bootloader
really accepts. This patch fills the gap and let add (just for YML
configuration) a list of variables that can be set and their flags. The
default policy is that all variables are accepted, exactly as it is done
in U-Boot, and check is done only if the list is found in the
configuration file.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>