Make libyaml optional
NewYAML format is required for extended features because the format
foreseen by U-Boot is very limited. However, some systems due to
low resources don't want to link to libyaml. Add the option
NO_YML_SUPPORT to disable YAML configuration file and just use
fw_env.config in the U-Boot format.

There are no functional changes in this patch - function depending on
YML are moved in a separate file, and some functions are factorized.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>